# swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "description": "Authorization API for STEXS",
        "title": "Authorization API",
        "version": "1.0.0"
    },
    "host": "localhost",
    "basePath": "/auth/api/v1/",
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "components": {
        "schemas": {
...
            "SuccessResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "A boolean indicator of the operation's success."
                    },
                    "message": {
                        "type": "string",
                        "description": "A human-readable message describing the result of the operation."
                    },
                    "timestamp": {
                        "type": "string",
                        "format":"ISO 8601",
                        "description": "The timestamp when the message was generated."
                    },
                    "data": {
                        "type": "object",
                        "description": "Additional relevant data returned with the response."
                    }
                },
                "example": {
                    "success": true,
                    "message": "Success message",
                    "timestamp": "2023-10-05T14:21:52.793Z",
                    "data": {}
                }
            }
        },
        "securitySchemes": {
            "Auth": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "Bearer Token"
            }
        }
    },
    "paths": {
...
        "/user": {
            "get": {
                "summary": "Retrieve User Data",
                "tags": ["User"],
                "description": "Allows authenticated users to retrieve their own user data. This endpoint provides access to essential user information, including email, raw user metadata, creation timestamp, update timestamp, and user ID.",
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Data Retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "description": "A unique UUID identifier for the user's account."
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "description": "The user's registered email address."
                                        },
                                        "raw_user_meta_data": {
                                            "type": "object",
                                            "description": "Raw user metadata associated with the user's account."
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "ISO 8601",
                                            "description": "Timestamp indicating when the user's account was created."
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "ISO 8601",
                                            "description": "Timestamp indicating when the user's account was last updated."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/password": {
            "post": {
                "summary": "Change Password",
                "tags": ["User"],
                "description": "Allows authenticated users to change their account password. Use this endpoint to securely update the user's password.",
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "description": "The new password to be set for the user's account."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password Change Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/email": {
            "post": {
                "summary": "Initialize Email Change",
                "tags": ["User"],
                "description": "Allows authenticated users to initialize a change to their registered email address. Use this endpoint to start the process of updating the email associated with the user's account.",
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "description": "The new email address to be associated with the user's account."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Email Change Initialization Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/email/verify": {
            "post": {
                "summary": "Verify Email Change",
                "tags": ["User"],
                "description": "Allows authenticated users to verify their email change by providing the verification code sent to their new email address. Use this endpoint to confirm the change of the email associated with the user's account.",
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "description": "The verification code sent to the user's new email address for email change confirmation."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Email Change Verification Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}